Refactor story validation into focused modules
